KYLIN: GRT: Fix compiling error in i386_defconfig
Mainline: KYLIN-only
Severity: Important

task: 449187

In 32bit arch, the 'resource_size_t' is defined as 'u32' in
"./include/linux/types.h" by 'CONFIG_ARCH_DMA_ADDR_T_64BIT'. There
exists a compiling error in 'grtnic_main.c' because using 'llu%'
to print 'u32'.

This patch use 'CONFIG_ARCH_DMA_ADDR_T_64BIT' to fix this.

K2CI-Arch: All
Reported-by: k2ci <kernel-bot@kylinos.cn>
Reviewed-on: http://gerrit.kylin.com/c/klinux/+/165017
Code-Style: k2ci <kernel-bot@kylinos.cn>
Tested-by: k2ci <kernel-bot@kylinos.cn>

+++ drivers/net/ethernet/grtnic_xgb/grtnic_main.c
@@ -234,7 +234,9 @@ static int grtnic_map_bars(struct grtnic_adapter *adapter, struct pci_dev *pdev)
 		dev_err(dev, "Could not map USER BAR");
 		return -1;
 	}
-	dev_info(dev, "USER BAR mapped at 0x%p with length %llu", hw->user_bar, bar_len);
+
+	dev_info(dev, "USER BAR mapped at 0x%p with length %llu",
+		 hw->user_bar, (unsigned long long)bar_len);
 
 	bar_start = pci_resource_start(pdev, 1);
 	bar_end = pci_resource_end(pdev, 1);
@@ -248,7 +250,8 @@ static int grtnic_map_bars(struct grtnic_adapter *adapter, struct pci_dev *pdev)
 		return -1;
 	}
 
-	dev_info(dev, "DMA BAR mapped at 0x%p with length %llu", hw->dma_bar, bar_len);
+	dev_info(dev, "DMA BAR mapped at 0x%p with length %llu",
+		 hw->dma_bar, (unsigned long long)bar_len);
 
 	return 0;
 }
